chore(server): inline queryOnce into CacheRegression.test.ts
