+++ packages/server/test/local-backend/CacheRegression.test.ts
@@ -6,12 +6,33 @@
  * `Clock.currentTimeMillis` opt-in still evicts.
  */
 
-import type { Ref } from "@confect/core";
+import { Ref } from "@confect/core";
 import { describe, expect, layer } from "@effect/vitest";
-import { Duration, Effect } from "effect";
+import type { ConvexHttpClient } from "convex/browser";
+import { Duration, Effect, Schema } from "effect";
 import refs from "./fixtures/confect/_generated/refs";
 import * as LocalBackend from "./LocalBackend";
-import { queryOnce } from "./queryOnce";
+
+class ConvexQueryError extends Schema.TaggedError<ConvexQueryError>()(
+  "ConvexQueryError",
+  { message: Schema.String },
+) {}
+
+/**
+ * Run a Confect query once over HTTP. Accepts a Confect `Ref` so the return
+ * type is inferred at the call site via `Ref.Returns<R>`.
+ */
+const queryOnce = <R extends Ref.AnyPublicQuery>(
+  client: ConvexHttpClient,
+  ref: R,
+  ...args: Ref.OptionalArgs<R>
+): Effect.Effect<Ref.Returns<R>, ConvexQueryError> =>
+  Effect.tryPromise({
+    try: () =>
+      client.query(Ref.getFunctionReference(ref), (args[0] ?? {}) as never),
+    catch: (error) =>
+      new ConvexQueryError({ message: `query failed: ${String(error)}` }),
+  });
 
 // `MAX_CACHE_AGE` plus 1s of slack for scheduling jitter.
 const SLEEP_PAST_CACHE = Duration.sum(LocalBackend.maxCacheAge, "1 second");
